tests/python/requirements.txt
allure-pytest==2.14.2
av==16.0.1
pytest==6.2.5
pytest-cases==3.8.6
pytest-timeout==2.1.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
requests==2.32.4
deepdiff==8.6.1
boto3==1.35.95
Pillow==10.3.0
python-dateutil==2.8.2
pyyaml==6.0.2
numpy==2.0.0

# TODO: update pytest to 7.0.0 and pytest-timeout to 2.3.1 (better debug in vscode)
